Display and update extra data
we display extra data fields in the form. we allow updating them. we
keep the type of data the same -- if we want to change data type, we
should use the console. this will allow rapid updates of extra data,
which helps when rollup is being used as a config store